Rearrange some things, moving dtos into a dto subfolder and plugin implementations to an impl subfolder.
