+++ src/escrow/dto/add.dto.ts
@@ -1,6 +1,6 @@
 import { IsString, IsNotEmptyObject, IsEnum } from 'class-validator';
 import { ApiProperty } from '@nestjs/swagger';
-import { PluginTypeEnum } from '../plugins/plugin.type.enum';
+import { PluginTypeEnum } from '../../plugins/plugin.type.enum';
 
 /**
  * DTO for the add endpoint
