+++ src/sms/sms.service.ts
@@ -6,8 +6,8 @@ import { ProtocolErrorCode } from 'protocol-common/protocol.errorcode';
 import { SecurityUtility } from 'protocol-common/security.utility';
 import { SmsOtp } from '../entity/sms.otp';
 import { SmsErrorCode } from './sms.errorcode';
-import { SmsFiltersDto } from './dtos/sms.filters.dto';
-import { SmsParamsDto } from './dtos/sms.params.dto';
+import { SmsFiltersDto } from './dto/sms.filters.dto';
+import { SmsParamsDto } from './dto/sms.params.dto';
 import { RateLimitService } from '../ratelimit/ratelimit.service';
 import { RateLimitBucket } from '../ratelimit/ratelimit.bucket';
 import { ISmsService } from '../remote/sms.service.interface';
